requirements.txt
torch>=2.0.0
numpy
pandas
pyyaml
opencv-python
sounddevice
soundfile
nilearn
nibabel
openai-whisper
